+++ core/src/graphics/particles/RandomizedParticle.cpp
@@ -63,7 +63,7 @@ void RandomizedParticle::tick(Amount<Unit::SECOND> timespan)
 
 	const auto distance = speed.to<Unit::NONE>(timespan).asStd();
 	const Amount<Unit::RADIAN> radians = direction;
-	shape.move(sf::Vector2f(distance * std::cosf(radians.asStd()), distance * std::sinf(radians.asStd())));
+	shape.move(sf::Vector2f(distance * std::cos(radians.asStd()), distance * std::sin(radians.asStd())));
 
 	auto directionIncrement = targetDirection - direction;
 	if (directionIncrement > 180.0_o)
